backend/requirements.txt
# FastAPI and Server
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1

# Database
motor==3.6.0
pymongo==4.10.1

# File Processing
PyPDF2==3.0.1
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3
python-pptx==1.0.2
Pillow==11.0.0

# AI Models
google-generativeai==0.8.3
openai==1.58.1
anthropic==0.42.0

# RAG and Embeddings
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
langchain==0.3.13
langchain-community==0.3.13

# Text Processing
markdown==3.7
beautifulsoup4==4.12.3
lxml==5.3.0

# Export
reportlab==4.2.5
markdown2==2.5.1

# Utilities
aiofiles==24.1.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.6.1
